Change Cool Mode from Frost-guard to Away
Since manual set points get reset by Netatmo api after a certain time, being able to trigger the away would allow you to make automations for when the house is empty on a work day or a long weekend. Away trigger remains in place until turned off.

Frost-guard is not very useful for automation, so it could be relegated to the Netatmo app itself.